requirements.txt
pandas
matplotlib
requests
tabula-py